move triggerBalance/Txn to TPP
