+++ user-sample/src/test/java/io/token/sample/NotifySampleTest.java
@@ -5,7 +5,6 @@ import static io.token.sample.TestUtil.createMemberAndLinkAccounts;
 import static io.token.sample.TestUtil.randomAlias;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.google.common.collect.ImmutableList;
 import io.token.proto.common.alias.AliasProtos.Alias;
 import io.token.proto.common.notification.NotificationProtos.NotifyStatus;
 import io.token.user.Member;
@@ -27,19 +26,6 @@ public class NotifySampleTest {
                     tokenClient,
                     payee,
                     payerAlias);
-            assertThat(status).isNotNull();
-        }
-    }
-
-    @Test
-    public void triggerBalanceStepUpNotificationTest() {
-        try (TokenClient tokenClient = createClient()) {
-            Member member = tokenClient.createMemberBlocking(randomAlias());
-
-            NotifyStatus status = member.triggerBalanceStepUpNotificationBlocking(
-                    ImmutableList.of(
-                            "123",
-                            "456"));
             assertThat(status).isNotNull();
         }
     }
